# Faster for the pour souls without symlinks