--- a/appveyor.yml
+++ b/appveyor.yml
@@ -4,9 +4,10 @@ version: '{build}'
 
 environment:
   matrix:
-    - nodejs_version: '2.0'
-    - nodejs_version: '1.8.1'
+    - nodejs_version: '0.10'
     - nodejs_version: '0.12'
+    - nodejs_version: '4.2'
+    - nodejs_version: '5.0'
 
 install:
   - ps: Install-Product node $env:nodejs_version