--- a/package.json
+++ b/package.json
@@ -9,11 +9,9 @@
     "type": "git",
     "url": "https://github.com/broccolijs/node-symlink-or-copy"
   },
-  "dependencies": {
-    "copy-dereference": "^1.0.0"
-  },
   "scripts": {
-    "test": "mocha tests/"
+    "test": "mocha tests/",
+    "test:debug": "mocha debug tests"
   },
   "devDependencies": {
     "mocha": "^2.2.4"