@stefanpenner 2016-04-04
@kellyselden / @raytiley if you have time to try this out (without symlinks enabled)