@joliss 2016-04-04
Looks good to me. @stefanpenner wanted to get some more testing in -- he'll merge and release once he feels it's ready.