@kellyselden 2016-04-05
### Initial without symlinks - before
```
Build successful - 63290ms.


Slowest Trees                                 | Total

----------------------------------------------+---------------------

Babel                                         | 11351ms

Babel                                         | 5700ms

TreeMerger (appAndDependencies)               | 3346ms

TreeMerger (stylesAndVendor)                  | 3319ms


Slowest Trees (cumulative)                    | Total (avg)

----------------------------------------------+---------------------

Babel (22)                                    | 25164ms (1143 ms)

TreeMerger (appAndDependencies) (1)           | 3346ms

TreeMerger (stylesAndVendor) (1)              | 3319ms
```

### Rebuild without symlinks - before
```
Build successful - 36673ms.


Slowest Trees                                 | Total

----------------------------------------------+---------------------

SimpleConcat: Concat: Test Support CSS        | 2495ms

SourceMapConcat: Concat: Addon JS             | 2454ms

StubGenerator                                 | 2200ms

Funnel: Funnel

  bower_components/ember-q... | 1980ms

SourceMapConcat: Concat: App                  | 1935ms


Slowest Trees (cumulative)                    | Total (avg)

----------------------------------------------+---------------------

BroccoliMergeTrees (9)                        | 3263ms (362 ms)

SimpleConcat: Concat: Test Support CSS (1)    | 2495ms

SourceMapConcat: Concat: Addon JS (1)         | 2454ms

StubGenerator (2)                             | 2245ms (1122 ms)

Funnel: Funnel

  bower_components/emb... (1) | 1980ms

SourceMapConcat: Concat: App (1)              | 1935ms
```
### Initial without symlinks - after
```
Build successful - 47696ms.


Slowest Trees                                 | Total

----------------------------------------------+---------------------

Babel                                         | 12116ms

Babel                                         | 4655ms

Babel                                         | 2864ms

Funnel                                        | 2617ms


Slowest Trees (cumulative)                    | Total (avg)

----------------------------------------------+---------------------

Babel (22)                                    | 26028ms (1183 ms)

BroccoliMergeTrees (9)                        | 3792ms (421 ms)

Funnel: Addon JS (16)                         | 3103ms (193 ms)

Funnel (6)                                    | 2658ms (443 ms)
```
### Rebuild without symlinks - after
```
Build successful - 2124ms.


Slowest Trees                                 | Total

----------------------------------------------+---------------------

BroccoliMergeTrees                            | 251ms


Slowest Trees (cumulative)                    | Total (avg)

----------------------------------------------+---------------------

BroccoliMergeTrees (9)                        | 365ms (40 ms)

Babel (22)                                    | 211ms (9 ms)
```