@raytiley 2016-04-05
I'm traveling till tomorrow but will run this against our app tomorrow evening. Looks sweet!