@ctataryn 2016-04-05
@stefanpenner you may want to update the the [ReadMe](https://github.com/broccolijs/node-symlink-or-copy#notes) about using junctions. Currently you have open questions about their usefulness.