@raytiley 2016-04-14
I can only blame myself for saying I would test this and dropping the ball, but after this got released last night our builds started failing. I'll post some more details in a bit.