@stefanpenner 2016-04-14
Strange, tested it in afew apps myself. But look forward to seeing more details.

Will rollback and release if needed