@stefanpenner 2016-04-15
@raytiley https://github.com/broccolijs/node-symlink-or-copy/pull/16