@ctataryn 2016-04-19
FYI: Apr 15th was the last successful build we had on Windows before encountering the following error:

> ENOENT: no such file or directory, stat 'E:\JW\Titles Online Folder\tpr-tol-ui\tmp\broccoli_merge_trees-input_base_path-KxqDUDjb.tmp\3\tpr-tol-ui\config'

> NPM version
> 2.14.12
> Ember version
> Future versions of Ember CLI will not support v4.2.4. Please update to Node 0.12 or io.js.
> version: 1.13.8
> node: 4.2.4
> npm: 2.13.4
> os: win32 x64
> Git Version
> git version 2.7.0.windows.1
> 

@stefanpenner when we had talked, you had suggested that it could be related to this patch, however you merged things on Apr 13th so to me the dates don't exactly align.  

This build is being done through Jenkins and when I inspect the `tmp` folder in the Workspace it is completely empty... Another note, this error doesn't actually cause the build to fail so I'm not sure if the proper exit level is being raised.  The entire log can be found here:

https://gist.github.com/ctataryn/d59f7df0b6c0ad72d425d84bd984d74c

@stefanpenner if you feel these aren't related I can open an issue at ember-cli. Let me know.