@ctataryn 2016-04-19
Ahh ok, I see that npm was only updated 2 days ago. So this could actually explain it...