@ctataryn 2016-04-19
Ok, solved my problem by:

1. running `npm cache clean`
1. adding an explicit dependency for `symlink-or-copy` at `1.1.0`
1. re-running my `ember build`

`tmp` folder is still empty after the build is done, perhaps the ember folks are [boy scouting](http://programmer.97things.oreilly.com/wiki/index.php/The_Boy_Scout_Rule) their `tmp` folders these days...