@stefanpenner 2016-04-20
thanks for verifying pining works. I have a reproduction of another app. But haven't had a chance to dig in, hope to soon. 

This was very thoroughly tested, although it seems like the category of unknown unknowns was larger then expected.